deploy v1777816754

+++ internal/pipeline/pipeline.go
@@ -7,6 +7,7 @@ import (
 	"os/exec"
 	"time"
 )
+
 type Step struct {
 	Name    string
 	Command string
@@ -15,37 +16,55 @@ type Step struct {
 func RunPipeline(repo string) {
 	log.Println("🚀 Pipeline started for:", repo)
 
-	// 🟢 version unique
+	// 🔥 version unique
 	version := fmt.Sprintf("v%d", time.Now().Unix())
+
+	// 🔐 token GitHub (env variable)
 	token := os.Getenv("GITHUB_TOKEN")
+
+	if token == "" {
+		log.Println("❌ GITHUB_TOKEN not set")
+		return
+	}
+
+	// 🐳 image Docker versionnée
 	imageName := "nourhenhachem/innodeploy-app:" + version
+
+	// 🔗 remote Git avec auth
+	remote := fmt.Sprintf(
+		"https://hachemn:%s@github.com/hachemn/innodeploy-pfe.git",
+		token,
+	)
 
 	steps := []Step{
 		{"Cleanup", "rm -rf project"},
 		{"Clone", "git clone " + repo + " project"},
-		// build
+
+		// 🐳 Build
 		{"Build Docker Image", "cd project && docker build -t innodeploy-app ."},
 
-		// tag avec version
+		// 🏷️ Tag version
 		{"Tag Image", "docker tag innodeploy-app " + imageName},
 
-		// push
+		// 🚀 Push Docker
 		{"Push Image", "docker push " + imageName},
 
-		// 🔥 update YAML (GitOps)
+		// 🧠 GitOps → update YAML
 		{"Update YAML", fmt.Sprintf(
 			"sed -i 's|image:.*|image: %s|' k8s/deployment.yaml",
 			imageName,
 		)},
-		{
- 		 "Set Git Remote",
- 		 fmt.Sprintf("git remote set-url origin https://%s@github.com/hachemn/innodeploy-pfe.git", token),
-		},
-		// push git
+
+		// 🔐 config remote avec token
+		{"Set Git Remote", "git remote set-url origin " + remote},
+
+		// 📝 commit
 		{"Git Commit", fmt.Sprintf(
 			"git add . && git commit -m 'deploy %s'",
 			version,
 		)},
+
+		// 📤 push
 		{"Git Push", "git push"},
 	}
